Refactor/#191 wishlist 오타 및 uri 수정 (#192)
* refactor: features wishlist uri 변경
* fix: wishilist 오타 수정
