+++ src/features/wishlist/api/deleteWishlist.ts
@@ -1,5 +1,5 @@
 import { deleteApi } from '@/shared';
 
 export default async function deleteWishlist(meetingId: number): Promise<void | null> {
-  return deleteApi<void>(`/api/v1/meetings/${meetingId}`);
+  return deleteApi<void>(`/api/v1/wishlist/${meetingId}`);
 }
