+++ src/features/wishlist/api/postWishlist.ts
@@ -1,5 +1,5 @@
 import { postApi } from '@/shared';
 
 export default async function postWishlist(meetingId: number): Promise<void | null> {
-  return postApi<void>(`/api/v1/meetings/${meetingId}`, {});
+  return postApi<void>(`/api/v1/wishlist/${meetingId}`, {});
 }
