+++ src/features/wishlist/model/hooks/usePostWishlistMutation.ts
@@ -1,5 +1,5 @@
 import { useMutation, UseMutationResult, useQueryClient } from '@tanstack/react-query';
-import wishlistKeys from '@/entities/wishilist/model/wishlist.keys';
+import wishlistKeys from '@/entities/wishlist/model/wishlist.keys';
 import postWishlist from '../../api/postWishlist';
 
 export default function usePostWishlistMutation(): UseMutationResult<
